Rename birthday subcommand files
