+++ src/commands/birthday/subcommands/index.ts
@@ -1,13 +1,13 @@
-import { list } from './_list';
-import { messageAdd } from './_messageAdd';
-import { messageDelete } from './_messageDelete';
-import { messageList } from './_messageList';
-import { scheduleGet } from './_scheduleGet';
-import { scheduleSet } from './_scheduleSet';
-import { set } from './_set';
-import { setup } from './_setup';
-import { show } from './_show';
-import { unset } from './_unset';
+import { list } from './list';
+import { messageAdd } from './messageAdd';
+import { messageDelete } from './messageDelete';
+import { messageList } from './messageList';
+import { scheduleGet } from './scheduleGet';
+import { scheduleSet } from './scheduleSet';
+import { set } from './set';
+import { setup } from './setup';
+import { show } from './show';
+import { unset } from './unset';
 
 export const subcommands = [
   list,
