+++ src/commands/birthday/subcommands/setup.ts
